requirements.txt
fastapi
uvicorn
azure-data-tables
python-multipart